pyproject.toml
[project]
name = "llm_orchestrator"
version = "0.1.1"
description = "Orchestrates LLM agent DAGs via YAML"
authors = [
  { name = "Abhijnan Acharya", email = "abhijnanacharya11@gmail.com" }
]
requires-python = ">=3.10"

dependencies = [
  "openai",
  "pyyaml",
  "python-dotenv",
  "schedule",
  "imaplib2",
  "requests",
  "beautifulsoup4",
  "typer"
]

[project.scripts]
talos = "cli.main:run_cli"

[tool.setuptools]
package-dir = {"" = "src"}

packages = [
  "cli",
  "email_client",
  "orchestrator",
  "orchestrator.agents",
  "utils"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
